front-end/components: Added component and dropdown support to texblock components in the front-end in log and right sidebar goals
