+++ src/front-end/components/block.css
@@ -1,7 +1,14 @@
 :host {
 	font-family:"SF-Pro", "roboto";
 	position: relative;
+	display: block;
 }
+
+:host(:hover) #editorIcons[tracker="true"]{
+	opacity: 1;
+	transition: 0.3ms ease-in-out;
+}
+
 #textBlock {
 	font-family: "SF-Pro";
 	border: none;
@@ -70,6 +77,31 @@
 	width: calc(100% - 38.3px - var(--block-tab-level));
 }
 
+.unstylized.tracker{
+	width: calc(100% - var(--block-tab-level)) !important;
+}
+
+.note.tracker{
+	margin-left: 0 !important;
+	width: calc(100% - var(--block-tab-level)) !important;
+}
+
+.task.tracker{
+	width: calc(100% - 23.3px) - var(--block-tab-level) !important;
+}
+
+.header1.tracker{
+	width: calc(100% - var(--block-tab-level)) !important;
+}
+
+.header2.tracker{
+	width: calc(100% - var(--block-tab-level)) !important;
+}
+
+.header3.tracker{
+	width: calc(100% - var(--block-tab-level)) !important;
+}
+
 .bold{
 	font-weight: 900;
 }
@@ -98,14 +130,6 @@
 	cursor: pointer;
 	filter: var(--icon-filter);
 }
-
-/* #editorIcons div{
-	display: inline-block;
-	margin-right: 5px;
-	opacity: 1;
-	width: 20px;
-	cursor: pointer;
-} */
 
 .unfocusedIcons{
 	opacity: 0.3;
@@ -142,6 +166,17 @@
 
 .paragraphIcons{
 	margin-top: 9px;
+}
+
+#editorIcons[tracker="true"] {
+	position: absolute;
+	top: calc(50% - 40px);
+	right: -25px;
+	padding-left: 15px;
+	padding-top: 10px;
+	padding-bottom: 10px;
+	opacity: 0;
+	transition: 0.3ms ease-in-out;
 }
 
 #checkerContainer{
@@ -246,4 +281,41 @@
 	top: -10px;
 	margin: 0 0 -7px;
 	width: calc(100% - 35px - var(--block-tab-level));
+}
+
+.paragraphSignifiers.tracker{
+	left: 0;
+	top: -12px;
+	margin: 0 0 -7px;
+	width: calc(100% - var(--block-tab-level));
+}
+.noteSignifiers.tracker{
+	left: 0;
+	top: 0;
+	margin: 0 0 -7px;
+	width: calc(100% - 13px - var(--block-tab-level));
+}
+.taskSignifiers.tracker{
+	left: 10px;
+	top: -5px;
+	margin: 0 0 -5px;
+	width: calc(100% - 10px - var(--block-tab-level));
+}
+.h1Signifiers.tracker{
+	left: 0;
+	top: -20px;
+	margin: 0 0 -7px;
+	width: calc(100% - var(--block-tab-level));
+}
+.h2Signifiers.tracker{
+	left: 0;
+	top: -13px;
+	margin: 0 0 -7px;
+	width: calc(100% - var(--block-tab-level));
+}
+.h3Signifiers.tracker{
+	left: 0;
+	top: -10px;
+	margin: 0 0 -7px;
+	width: calc(100% - var(--block-tab-level));
 }
