+++ src/front-end/components/block.jsx
@@ -12,7 +12,7 @@ import { bindDropdown } from "./blockDropdownManager.js";
 /** @jsx createElement */
 /** @jsxFrag createFragment */
 import { createElement, createFragment } from "../jsxEngine.js";
-import { contentWrapper } from "../index.js";
+import "./shadow.js";
 /* eslint-enable */
 
 export let tabSize = 40;
@@ -25,6 +25,10 @@ let blockTemplate = <template>
 	<div id="textBlock" contentEditable="true" ondrop={()=>false} placeholder='Type "/" to create a block'></div>
 	<aside id="signifiers" class="paragraphSignifiers"><span>&#128540;</span><button>+</button></aside>
 </template>;
+
+let trackerIcons = <template>
+	<div id="editorIcons" tracker="true"><img src="../public/resources/sixDotIcon.png" class="unfocusedIcons" id="more" /></div>
+</template>
 
 /**
  * Class to create new editor block
@@ -40,7 +44,6 @@ export class TextBlock extends HTMLElement {
 	 */
 	constructor (controller, itemObject, signifiers, callback, tracker=null) {
 		super();
-		this.paddingSize = contentWrapper.getClientRects()[0].x + 38;
 		this.attachShadow({ mode: "open" });
 		this.shadowRoot.appendChild(blockTemplate.content.cloneNode(true));
 		this.characterIndex = 0;
@@ -48,6 +51,7 @@ export class TextBlock extends HTMLElement {
 		this.initialHeight = 3;
 		this.item = itemObject;
 		this.controller = controller;
+		this.paddingSize = this.controller.container.getClientRects()[0].x + 38;
 		this.shadowRoot.getElementById("textBlock").classList.add("unstylized");
 		this.currentBlock = null;
 		this.checkBox = this.shadowRoot.getElementById("checkerContainer");
@@ -56,8 +60,14 @@ export class TextBlock extends HTMLElement {
 		this.editorIcons = this.shadowRoot.getElementById("editorIcons");
 		this.signifierRow = this.shadowRoot.getElementById("signifiers");
 		this.currentPointerHeight = 2;
+		this.tracker = tracker;
+		if (this.tracker) {
+			this.shadowRoot.getElementById("editorIcons").remove();
+			this.signifierRow.insertAdjacentHTML("beforebegin", trackerIcons.innerHTML);
+			this.signifierRow.classList.add("tracker");
+			this.shadowRoot.getElementById("textBlock").classList.add("tracker");
+		}
 		bindFunctions(this);
-		bindDropdown(this);
 		if (this.controller.creatingFromBullet.isTrue) {
 			if (this.controller.creatingFromBullet.kind === "note") {
 				this.setupNote();
@@ -67,16 +77,18 @@ export class TextBlock extends HTMLElement {
 				this.setupTask();
 			}
 		}
-		this.tracker = tracker;
+		
 		this.tabLevel = controller.currentTabLevel;
 		this.atPressed = false;
 		this.hashPressed = false;
 		this.timeSetter = false;
 		this.dateSetter = false;
 		this.eventDelete = true;
-		this.signifiers = signifiers
+		this.signifiers = signifiers;
+		
 		this.plus = this.shadowRoot.getElementById("plus");
 		this.more = this.shadowRoot.getElementById("more");
+		bindDropdown(this);
 		this.setupTabLevel();
 		setTimeout(() => {
 			callback(true);
@@ -97,94 +109,3 @@ export class TextBlock extends HTMLElement {
 
 window.customElements.define("text-block", TextBlock);
 
-
-const SUPPORTS_SHADOW_SELECTION = typeof window.ShadowRoot.prototype.getSelection === 'function';
-const SUPPORTS_BEFORE_INPUT = typeof window.InputEvent.prototype.getTargetRanges === 'function';
-const IS_FIREFOX = window.navigator.userAgent.toLowerCase().indexOf('firefox') > -1;
-
-class ShadowSelection {
-  constructor() {
-    this._ranges = [];
-  }
-
-  getRangeAt(index) {
-    return this._ranges[index];
-  }
-
-  addRange(range) {
-    this._ranges.push(range);
-  }
-
-  removeAllRanges() {
-    this._ranges = [];
-  }
-
-  // todo: implement remaining `Selection` methods and properties.
-}
-
-function getActiveElement() {
-  let active = document.activeElement;
-
-  while (true) {
-    if (active && active.shadowRoot && active.shadowRoot.activeElement) {
-      active = active.shadowRoot.activeElement;
-    } else {
-      break;
-    }
-  }
-
-  return active;
-}
-
-if (IS_FIREFOX && !SUPPORTS_SHADOW_SELECTION) {
-  window.ShadowRoot.prototype.getSelection = function() {
-    return document.getSelection();
-  }
-}
-
-if (!IS_FIREFOX && !SUPPORTS_SHADOW_SELECTION && SUPPORTS_BEFORE_INPUT) {
-  let processing = false;
-  let selection = new ShadowSelection();
-
-  window.ShadowRoot.prototype.getSelection = function() {
-    return selection;
-  }
-
-  window.addEventListener('selectionchange', () => {
-    if (!processing) {
-      processing = true;
-
-      const active = getActiveElement();
-
-      if (active && (active.getAttribute('contenteditable') === 'true')) {
-        document.execCommand('indent');
-      } else {
-        selection.removeAllRanges();
-      }
-
-      processing = false;
-    }
-  }, true);
-
-  window.addEventListener('beforeinput', (event) => {
-    if (processing) {
-      const ranges = event.getTargetRanges();
-      const range = ranges[0];
-
-      const newRange = new Range();
-
-      newRange.setStart(range.startContainer, range.startOffset);
-      newRange.setEnd(range.endContainer, range.endOffset);
-
-      selection.removeAllRanges();
-      selection.addRange(newRange);
-
-      event.preventDefault();
-      event.stopImmediatePropagation();
-    }
-  }, true);
-
-  window.addEventListener('selectstart', (event) => {
-    selection.removeAllRanges();
-  }, true);
-}
